requirements_test.txt
# Test dependencies for Fallback Sensors integration
pytest>=7.4.0
pytest-homeassistant-custom-component>=0.13.0
